Add files via upload
This is a self assessment project I am working on as an internship project. It involves analyzing retail data to show wha tareas have the most losses or least profit as well as what product is doing best in terms of sales in the USA. I used Tableau for my visualizations.